requirements.txt
asgiref
bcrypt
certifi
cffi
charset-normalizer
cryptography
Django
django-cors-headers
djangorestframework
idna
librouteros
mysqlclient
paramiko
Pillow
pycparser
PyNaCl
pytz
requests
sqlparse
urllib3
